refactor: move test to functional, add optional chaining and cockroachdb driver
