+++ test/functional/relations/bigint-join-column/entity/Category.ts
@@ -1,4 +1,4 @@
-import { Column, Entity, PrimaryColumn } from "../../../../src"
+import { Column, Entity, PrimaryColumn } from "../../../../../src"
 
 @Entity()
 export class Category {
