+++ test/functional/relations/bigint-join-column/entity/Post.ts
@@ -1,4 +1,9 @@
-import { Entity, PrimaryGeneratedColumn, ManyToOne, JoinColumn } from "../../../../src"
+import {
+    Entity,
+    JoinColumn,
+    ManyToOne,
+    PrimaryGeneratedColumn,
+} from "../../../../../src"
 import { Category } from "./Category"
 
 @Entity()
